Moved InputFields to separate folder under Utils
