+++ src/App/App.scss
@@ -18,7 +18,7 @@ html, body, p, div, h1, h2, h3, h4, h5 {
 }
 
 a, p, h1, h2, h3, h4, h5, span, a:hover {
-  color: black;
+  color: $COLOR-font-dark;
   text-decoration: none;
 }
